Rename Headings directory to 'Typography' and move NoteText component. Update all imports accordingly.
