+++ src/Typography/SectionHeading.tsx
