+++ src/Typography/index.ts
@@ -1,2 +1,3 @@
 export { default as PageTitle } from './PageTitle';
 export { default as SectionHeading } from './SectionHeading';
+export { default as NoteText } from './NoteText';
